refactor: move constructChatMessage function into a util file
